NMBasicModel-class(RNMImport)	R Documentation

Standard/'basic' NONMEM individual problem

Description

This class holds the results from an individual NONMEM problem (not an entire control file) that has no $SIM step.

Objects from the Class

Individual objects of this class are not meant to be instantiated on their own, but are created when loading a NONMEM run via importNm

Slots

parameterIterations:

A data.frame of the iteration of each parameter estimate, if available

objectiveFinal:

The numeric value of the objective function minimum

thetaFinal:

Final estimates of the "thetas", together with the standard errors, if available (as a 2-row matrix))

sigmaFinal:

Final estimates of the "sigmas", together with the standard errors, if available (as an array with 1 or 2 matrices)

omegaFinal:

Final estimates of the "omegas", together with the standard errors, if available (as an array with 1 or 2 matrices)

parameterCovMatrix:

The variance-covariance of the parameter estimators, if available

minInfo:

A string describing the status of the objective function-minimization

problemStatement:

A string holding the $PROBLEM statement

controlStatements:

List of parsed sections from the control file

lstStatements:

List of parsed sections from the list file

inputData:

A data.frame holding the run's input data. If it could not be read for some reason, it will be empty

outputData:

A data.frame holding the run's aggregated output data, (from the files in the $TABLE statement)

additionalVars:

A data.frame of user-created variables (usually categorical variables created through addDerivedCategorical

Extends

The virtual class "NMProblem", directly.

Methods

getOmegas

signature(obj = "NMBasicModel"): Retrieves the omega estimates from the problem

getSigmas

signature(obj = "NMBasicModel"): Retrieves the sigma estimates from the problem

getThetas

signature(obj = "NMBasicModel"): Retrieves the theta estimates from the problem

addDerivedCategorical

signature(obj = "NMBasicModel"): Create user-defined categorical variables

imposeCategoryFormat

signature(obj = "NMBasicModel"): Coerces variables to categories

nmData

Extracts the input and output data from the problem

show

getObjective

signature(obj = "NMBasicModel")

getEstimateCov

signature(obj = "NMBasicModel"): Extracts the covariance matrix and optionally the correlation matrix of the estimators for this problem.

Author(s)

Mango Solutions <support@mango-solutions.com>

[Package RNMImport version 0.1.9 Index]