NMRun-class(RNMImport)	R Documentation

NONMEM run container class

Description

This is the basic class for handling a standard NONMEM run. In essence, it holds several individual problems which may be of type NMBasicModel, NMSimModel, and NMSimDataGen. It also holds the control file and list output file text

Objects from the Class

Objects of this class are typically created with the importNm function, which imports a NONMEM run on the basis of its control file.

Slots

controlText:

Text of the control file, without comments

lstText:

Text of the output "list" file

controlComments:

Comments of each line of the control file

controlFileInfo:

Information about the control file

listFileInfo:

Information about the list file

numProblems:

Number of problems in the run

problems:

List of the actual problem results

Methods

getOmegas

signature(obj = "NMRun"): Retrieves the omega estimates from a particular problem

getSigmas

signature(obj = "NMRun"): Retrieves the sigma estimates from a particular problem

getThetas

signature(obj = "NMRun"): Retrieves the theta estimates from a particular problem

addDerivedCategorical

signature(obj = "NMRun"): Create user-defined categorical variables

imposeCategoryFormat

signature(obj = "NMRun"): Coerces variables to categories

nmData

Extracts the input and/or output data of one of the problems in the run.

show

Prints basic information about the run.

getEstimateCov

signature(obj = "NMRun"): ...

addedData

signature(obj = "NMRun"): Extracts user-created data (typically through) addDerivedCategorical into a data.frame from a specified problem.

getObjective

signature(obj = "NMRun")

getControlStatements

signature(obj = "NMRun"): Extract an object with parsed statements control file corresponding to a particular problem

Author(s)

Mango Solutions <support@mango-solutions.com>

Examples


         showClass("NMRun")

[Package RNMImport version 0.1.9 Index]