NMSimDataGen-class(RNMImport)	R Documentation

Class for holding simulation problems that don't fit models

Description

This class handles NONMEM models that have a $SIM statement but only simulate data without fitting a mode

Objects from the Class

Objects of this class are not usually instantiated on their own, but are created as part of a call to importNm.

Slots

numSimulations:

Number of simulations/subproblems

seeds:

Seeds used for the random number generator

thetaInitial:

thetaInitial Fixed/initial theta values used to generate data (numeric vector)

omegaInitial:

Fixed/initial omega values used to generate data (numeric matrix)

sigmaInitial:

Fixed/initial sigma values used to generate data (numeric matrix)

problemStatement:

The contents of the $PROB statement

controlStatements:

A list of parsed sections of the control file

lstStatements:

A list of parsed sections of the output "list" file

inputData:

A data.frame of the input data, if available (otherwise an empty data.frame)

outputData:

Aggregation of the output data

additionalVars:

A data.frame of additional variables created by the user

Extends

Virtual class "NMProblem", directly.

Methods

addDerivedCategorical

signature(obj = "NMSimDataGen"): Create user-defined categorical variables

imposeCategoryFormat

signature(obj = "NMSimDataGen"): Coerces variables to categories

nmData

signature(object = "NMSimDataGen"): Retrieves the input/output data, where the output can come from 1 or more subproblems.

show

signature(object = "NMSimDataGen"): Prints basic data about the problem

Author(s)

Mango Solutions <support@mango-solutions.com>

Examples


        showClass("NMSimDataGen")

[Package RNMImport version 0.1.9 Index]