getNmFileExtensions(RNMImport)	R Documentation

Gets allowed NONMEM file extensions

Description

Returns a vector of allowed file extensions for a specific NONMEM file type, e.g. control, list, output table or input data.

Usage


getNmFileExtensions(parameter)

Arguments

parameter 	 The file type. Must be one of control, list, table or input

Value

A vector of appropriate file extensions

Author(s)

Mango Solutions <rweeks@mango-solutions.com>

[Package RNMImport version 0.1.5 Index]