getVarDescription(RNMImport)	R Documentation

Gets a variable description

Description

Gets the label, type and format of a selection of user specified variables, e.g. SEX and IPRED.

Usage


getVarDescription(varName)

Arguments

varName 	 A vector of variable names

Value

A data frame of required variable names, labels, types and formats

Author(s)

Mango Solutions <rweeks@mango-solutions.com>

[Package RNMImport version 0.1.5 Index]