importNm(RNMImport)	R Documentation

Import a NONMEM run

Description

Imports a NONMEM run based on a control file and an output list file. This function assumes that all of the input and output data tables are in the same directory as the control file, though the input tables are allowed to be missing.

Usage

importNm(conFile, reportFile = NULL, path = NULL, dropInputColumns = TRUE)

Arguments

conFile 	 [C,1] control file name

reportFile 	 [C,1] The name of the list file. In the future this parameter will be optional, in which case importNm will attempt' to deduce the file name on its own based on the name of the conFile and the list file extensions stored in the meta-data.

path 	 The path where the files are located. Can be the name of a path stored with setNmPath if it is surrounded by round brackets

dropInputColumns

{Logical flag. If TRUE, those columns of the input data flagged for dropping in the $INPUT statement will not be imported with the run}

Value

An NMRun object that holds the information of the NONMEM run

Author(s)

Mango Solutions <support@mango-solutions.com>

Examples


        x <- importNm("TestData1.ctl", "TestData1.lst", path  = system.file(package = "RNMImport", "unittests/testdata/TestRun"))
        print(x)

[Package RNMImport version 0.1.9 Index]