importNmReport(RNMImport)	R Documentation

Import the information in an output report file

Description

This routine imports the contents in a NONMEM output report file, and then parses different sections of it while ignoring others.

Usage

importNmReport(fileName, path, controlStatements)

Arguments

fileName 	 Name of the list file

path 	 (optional) Path to the list file. Can be a path stored by setNmPath if it is in round brackets.

controlStatements 	 (optional) Control statements from the run associated with this output. Using it allows cleaner parsing of the report file

Value

A named list with elements holding parsed statements in the list file

Author(s)

Mango Solutions

Examples


        x <- importNmReport("TestData1.lst", path = system.file(package = "RNMImport", "unittests/testdata/TestRun"))
        print(x)


[Package RNMImport version 0.1.6 Index]