importNmMod(RNMImport)	R Documentation

Read a NONMEM control file

Description

Parses a NONMEM control file, and returns its contents as a list of parsed elements. These elements will correspond to the actual control statements, e.g. $PK, $THETA, $PROBLEM, etc. Some of the elements are kept as pure text.

Usage

importNmMod(fileName, path)

Arguments

fileName 	 Name of the control file

path 	 (optional) path to the control file, can be a stored path enclosed in round brackets

Value

A list with parsed components of the control file

Author(s)

Mango Solutions

Examples


        conFile <- system.file(package = "RNMImport", "unittests/testdata/control4.con")
        conContents <- importNmMod(file = conFile)
        print(conContents)

[Package RNMImport version 0.1.6 Index]