nmData(RNMImport)	R Documentation

Extract input and / or output data tables from a NONMEM object

Description

A generic function that extracts input and output data tables from a NONMEM object into either a single consolidated data.frame or a list.

Usage


nmData(obj, dataTypes = c("input", "output"), returnMode = c("singleDF", "DFList"), ...)

Arguments

obj 	 An object of class NMRun, or one that inherits from NMProblem

dataTypes 	 Which type of data should be returned, must be "input" and/or "output"

returnMode 	 Which type of data should be returned, must be "input" and/or "output". If "singleDF", attempts to return a consolidated data.frame of output. and input data, if "DFList" returns a list with separate input and output data

... 	 Additional arguments that apply to different classes

Value

data.frame or list with "input" and "output" components

Author(s)

Mango Solutions <fgochez@mango-solutions.com>

Examples


        x <- importNm("TestData1.ctl", "TestData1.lst", path  = system.file(package = "RNMImport", "unittests/testdata/TestRun"))
        x.data <- nmData(x)
        print(head(x.data))
        x.data <- nmData(x, returnMode = "DFList" )
        print(lapply(x.data, head))
        y <- importNm("TestData1SIM.con", "TestData1SIM.lst", path  = system.file(package = "RNMImport", "unittests/testdata/TestSimRun"))
        y.outputData <- nmData(y, subProblemNum = 2:4, dataTypes = "output")
        print(summary(y.outputData))
        

[Package RNMImport version 0.1.6 Index]