setNmFileExtensions(RNMImport)	R Documentation

Sets NONMEM file extensions

Description

Sets new NONMEM file extensions for specific file types, e.g. one may want to change the allowed extensions for a NONMEM list file.

Usage


setNmFileExtensions(parameter, extension)

Arguments

parameter 	 The name of the NONMEM file type. Must be one of control, list, table or input

extension 	 A vector of new file extensions.

Value

None

Author(s)

Mango Solutions <rweeks@mango-solutions.com>

[Package RNMImport version 0.1.5 Index]