setVarDescription(RNMImport)	R Documentation

Sets new variable characteristics

Description

Sets a new label, format and type for an existing variable name, e.g. one may want to change the label for SEX or change the format of IPRED.

Usage


setVarDescription(varName, varLabel = NULL, varFormat = NULL, varType = NULL)

Arguments

varName 	 The variable name.

varLabel 	 The new label. NULL by default

varFormat 	 The new format. NULL by default

varType 	 The new type. NULL by default

Value

None

Author(s)

Mango Solutions <rweeks@mango-solutions.com>

[Package RNMImport version 0.1.5 Index]